# Speed up tests by enabling cache_classes