--- a/config/environments/test.rb
+++ b/config/environments/test.rb
@@ -11,7 +11,7 @@ Rails.application.configure do
   #
   # More details:
   # https://guides.rubyonrails.org/caching_with_rails.html#activesupport-cache-memcachestore
-  config.cache_classes = false
+  config.cache_classes = true
   config.action_view.cache_template_loading = true
   config.cache_store = :memory_store